Create user view page.

+++ src/maasserver/models.py
@@ -14,6 +14,7 @@ __all__ = [
     "NODE_STATUS",
     "Node",
     "MACAddress",
+    "UserProfile",
     ]
 
 import datetime
@@ -27,7 +28,10 @@ from django.core.files.base import ContentFile
 from django.db import models
 from django.db.models.signals import post_save
 from django.shortcuts import get_object_or_404
-from maasserver.exceptions import PermissionDenied
+from maasserver.exceptions import (
+    CannotDeleteUserException,
+    PermissionDenied,
+    )
 from maasserver.macaddress import MACAddressField
 from piston.models import (
     Consumer,
@@ -383,6 +387,26 @@ class MACAddress(CommonInfo):
 GENERIC_CONSUMER = 'Maas consumer'
 
 
+class UserProfileManager(models.Manager):
+    """A utility to manage the collection of UserProfile (or User).
+
+    This should be used when dealing with UserProfiles or Users because it
+    deals gracefully with system users.
+    """
+
+    def all_users(self):
+        """Returns all the "real" users (the users which are not system users
+        and thus have a UserProfile object attached to them).
+
+        :return: A QuerySet of the users.
+        :rtype: django.db.models.query.QuerySet_
+
+        """
+        userprofile_ids = UserProfile.objects.filter(
+            ).values_list('id', flat=True)
+        return User.objects.filter(id__in=userprofile_ids)
+
+
 class UserProfile(models.Model):
     """A User profile to store Maas specific methods and fields.
 
@@ -394,13 +418,20 @@ class UserProfile(models.Model):
 
     """
 
+    objects = UserProfileManager()
     user = models.OneToOneField(User)
 
     def delete(self):
+        if self.user.node_set.exists():
+            nb_nodes = self.user.node_set.count()
+            msg = (
+                "User %s cannot be deleted: it still has %d node(s) "
+                "deployed." % (self.user.username, nb_nodes))
+            raise CannotDeleteUserException(msg)
         self.user.consumers.all().delete()
         self.user.delete()
         super(UserProfile, self).delete()
-        
+
     def get_authorisation_tokens(self):
         """Fetches all the user's OAuth tokens.
